requirements.txt
streamlit
spacy
openai-whisper
torch==2.0.1
numpy
